Move 'application_builder' under 'application' package
